build: Updating dependencies
composer:
* mediawiki/mediawiki-codesniffer: 49.0.0 → 50.0.0
* mediawiki/mediawiki-phan-config: 0.18.0 → 0.20.0

npm:
* basic-ftp: 5.0.5 → 5.2.0
  * https://github.com/advisories/GHSA-5rq4-664w-9x2c
* fast-xml-parser: 5.3.7 → 5.5.3
  * https://github.com/advisories/GHSA-fj3w-jwp8-x2g3

+++ includes/formats/CargoCategoryFormat.php
@@ -46,7 +46,7 @@ class CargoCategoryFormat extends CargoListFormat {
 		$num = count( $valuesTable );
 
 		$prev_first_char = "";
-		$rows_per_column = ceil( $num / $numColumns );
+		$rows_per_column = (int)ceil( $num / $numColumns );
 		// Column width is a percentage.
 		$column_width = floor( 100 / $numColumns );
